arrays: fill GX pipe in arrays.cpp
Move the vertex format setup to the array reader classes, as well as
writing into the GX pipe. This not only buys us a little of performance,
but having the array reading code next to the pipe writing allows us to
implement improvements and new functionality more easily: for example,
the SameTypeVertexReader classes store the vertex data as bytes of
shorts into the GX pipe, if this is the data format specified by the
client.

This is also a step in the direction of implementing vertex buffers,
since we should now be able to add a VertexReader subclass that handles
indexed arrays (using GX_SetArray()).