TEV: use material color when no vertex color is given
Further simplify the rendering of non-lit objects: instead of storing
the constant color (in the case when the color is not given as vertex
data) in a TEV register, store it as material color: this allows us to
simplify the code a bit, because then we can hardcode both the color
channel, and set the TEV input to be the raster color.

Incidentally, we also remove an unused call to GX_SetChanCtrl(), because
we are not using the GX_COLOR1A1 channel (we have just a color channel
in this case).