TEV: implement texture environment GL_COMBINE operations
Since the implementation of GL_COMBINE is quite verbose, move the
texture setup code to a separate file (texture_unit.c).
We implement most of the GL_COMBINE functions, but due to hardware
limitations we cannot support the cases where more than one operand is
complemented (by "complemented" we mean the expression "1 - value");
these *could* be implemented using more stages, but let's not go into
that road for now.

The functions GL_DOT3_RGB and GL_DOT3_RGBA are not implemented, and
likely will never be, since the TEV does not provide a way to offset all
color operands by -0.5.

Note that for the time being the texture coordinate generation is
hardcoded to work on the first texture unit only; we'll fix it in
another commit, as this one is already white big.