Dynamically allocate GPU resources
Add a stack of structures holding the number of used GPU resources; this
is simpler than passing parameters around, and in any instant we can
know the allocation status.
At the moment the *_end members are not used, but they will come handy
once we implement some kind of shader support.