lighting: alpha should be multiplied, not summed
In the second color channel, the alpha component of the diffuse material
should not be added, but multiplied with the input from the previous
stage, because the OpenGL spec says that the value of alpha should be
taken from the alpha component of the color material's diffuse value.

This fixes the rendering of glass parts in Neverball.