state: do not rebuild TEV stages unless needed
Add a dirty flag to track whether the TEV stages must be rebuilt. The
replaces the dirty_lighting, dirty_clip_planes, dirty_stencil and the
(de facto unused) dirty_material flags, because all these factors
ultimately require the TEV to be rebuilt.

This is not super-precide, and we err on the side of rendering safety,
but it still brings a huge speed up in the crack-attack game, which uses
call lists with several draw commands in each (so the state is
guaranteed to be clean between one command and the next).