arrays: build vertex attributes readers at drawing time
Do not prepare the vertex attribute readers in advance. Instead, just
store the client data in the state and build the readers just at the
time when they are needed (like in glDrawArrays() and glDrawElements).
Keep a "dirty" flag to avoid rebuilding them if the input data has not
changed.

This refactoring is a small step toward supporting arbitrary vertex
attributes (needed for OpenGL 2.0+) and also allows implementing the
getter methods more easily (see next commit).

We also fix glInterleavedArrays(), which got broken at some point.