arrays: remove attribute readers from the global state
As the comment was saying, these objects are just for optimization
purposes and do not really belong to the global state variable. Add a
function to array.h so that other parts of the code can retrieve them if
needed, and move the actual variables as static variables in array.c.