examples: add OpenGL 2.0 texturing example
This example uses SDL2 and glm. At the moment this example does not work
on the Wii/GameCube, as the textured rendering using OpenGL 2.0 will
come in a later commit.

The "common" directory contains textures that get embedded into the
executable, so that examples can be written without access to external
files (which makes testing on the console -- or on the emulator --
very straightforward).