shader: add a client callback function for updating matrices
The client can update the matrices in the setup_draw() callback as well,
but having a specialized callback makes it possible for opengx to
request the client to update the matrices without setting up the whole
TEV state.
This is required for supporting stencil drawing when using shaders:
since the drawing happens in two stages (first we draw to the stencil
buffer, and then to the scene) and only the latter stage uses the TEV as
setup by the client, it would be a waste to call setup_draw() twice, if
all what we need from the first invocation is just the updated
transformation matrices.