shader: call the matrix update function via a function pointer
This allows to reduce branching, since the function pointer only changes
when glUseProgram() is called. We should consider using the same
mechanism for the setup_draw() function.

We also clean up the dirty bit usage for the normal matrix, partially
reverting commit 836fd7e7e8d45a5fc8232eed95b5164568754331, because the
current condition for rebuilding the normal matrix is anyway satisfied
when the dirty_matrices flag is set, so we can just set it anytime a
matrix is changed. We still keep setting the dirty_tev flag to true when
the normal matrix gets changed, in order not to break the idea behind
commit 836fd7e7e8d45a5fc8232eed95b5164568754331.